cast shape to int

+++ vgg16.py
@@ -291,10 +291,10 @@ class Vgg16:
         spatial_max = tf.reduce_max(activations, axis=[0, 1, 2])
 
         depth_argmax = tf.cast(tf.argmax(spatial_max, axis=-1), tf.int32)
-        spatial_argmax = tf.cast(utils.argmax_2d(activations)[0, :, depth_argmax], tf.int32)
+        spatial_argmax = utils.argmax_2d(activations)[0, :, depth_argmax]
 
         mask = tf.zeros_like(activations)
-        delta = tf.SparseTensor(spatial_argmax, [1.0], tf.shape(activations))
+        delta = tf.SparseTensor(spatial_argmax, [1.0], tf.cast(tf.shape(activations), tf.int32))
         mask += tf.sparse_tensor_to_dense(delta)
 
         activations *= mask
